requirements.txt
Django==2.2.5
djangorestframework==3.10.3
requests==2.22.0
psycopg2-binary==2.8.3
